+++ src/main/kotlin/it/polito/wa2group8/wallet/dto/WalletDTO.kt
@@ -5,7 +5,7 @@ import it.polito.wa2group8.wallet.domain.Wallet
 import java.math.BigDecimal
 
 data class WalletDTO(val customerId : Long?,
-                     val currentAmount : Long)
+                     val currentAmount : BigDecimal)
 
 
 fun Wallet.toWalletDTO() = WalletDTO(customer.customerID, currentAmount)
